Provide own mouse tracking for hyperlink implementation.
The previous method of making "actual" modifications to the content of the QTextDocument in order to implement hyperlinks was *wrong* so now we stick to using the "additional" formats in the text layout objects of the blocks int the document, and manually implement mouse tracking, tooltips etc. in the QTextEdit subclass. This also allows us to implement hyperlink clicking only when Ctrl (or Command) is pressed, or whatever.