Sync settings after each update. Make sure the settings dialog is up-to-date.
Make the settings dialog modal (so that you wouldn't be able to have
uncommitted changes in the dialog while making other changes in the main
window, which would then have to be "synchronized" with the settings dialog
while keeping the uncommitted changes intact). Update the settings dialog UI
every time it's shown in order to sync with the changes made in the main
window (i.e. font size).