Replace peg-markdown with multimarkdown.
The former depends on glib-2.0 (and of course I did not realize that the binary I added here was dynamically linked), and statically linking to glib-2.0 is apparently a bit nontrivial. We are bundling the peg-markdown dependencies for the Windows build and it seems to work okay, but multimarkdown has done away with the glib dependency, so we'll just go ahead and use that instead for OS X (and later Linux, and maybe also Windows).