Reimplement commitData() and saveState() in QarkdownApplication.
This is an attempt to get the "save changes?" prompt on quitting to work. On OS X these methods are not being called, and the signals that we attach in MainWindow are not being sent to its slots even though we connect them. Committing this so that I could test on Linux and Windows.