Perform "startup tasks" after the NSApplicationDidFinishLaunching event on OS X.
This fixes the issue where we would first open the last file that was open when MainWindow is created, and then another file that was requested via the "open file" event (e.g. by dropping the file on the application icon). Now we open the last file that was open *after* the app has finished launching (and handling the "open file" events), if a file is not already open.